requirement.txt
streamlit
python-docx
fpdf
pillow
PyMuPDF